# Load sound effects on a separate thread