--- a/CS483/CS483/Kartaclysm/Data/DevConfig/Preload.xml
+++ b/CS483/CS483/Kartaclysm/Data/DevConfig/Preload.xml
@@ -80,6 +80,7 @@
     <Sprite path="Assets/Hud/Positions/position_4.mtl" mtl="position_4"/>
 
     <!-- Menus -->
+    <Sprite path="Assets/Menus/Background/background.mtl" mtl="background"/>
     <Sprite path="Assets/Menus/MainMenu/loading_message.mtl" mtl="loading_message"/>
     <Sprite path="Assets/Menus/MainMenu/press_start.mtl" mtl="press_start"/>
     <Sprite path="Assets/Menus/MainMenu/title_image.mtl" mtl="title_image"/>