Surprisingly quick fix to load times. Doesn't work for ModelManager because of threading complications with the OpenGL state, but works fine with sound effects. 

For my few benchmark tests on VS 2015, it reduced the load time of this block by 25%
Release: 5s -> 3s
Debug: 15s -> 12s